DeBaser/requirements.txt
appdirs==1.4.3
Django==1.10.5
django-appconf==1.0.2
django-cors-headers==2.1.0
Django-Select2==5.8.10
djangorestframework==3.6.3
ez-setup==0.9
httpie==0.9.9
jsonfield==2.0.2
mysqlclient==1.3.10
packaging==16.8
pkg-resources==0.0.0
Pygments==2.2.0
PyMySQL==0.7.11
pyparsing==2.2.0
python-memcached==1.58
requests==2.14.2
requests-toolbelt==0.8.0
six==1.10.0
urllib3==1.22